chore: move `BitSet` to `rolldown_utils`
